+++ R/distcomp.R
@@ -63,7 +63,7 @@ NULL
 #' @rdname distcomp-internal
 #' @import utils
 #' @importFrom stringr str_trim
-#' @importFrom httr POST
+#' @importFrom httr POST stop_for_status
 #' @importFrom httr headers
 #' @importFrom httr add_headers
 #' @importFrom jsonlite fromJSON
@@ -115,6 +115,7 @@ NULL
 #' @param defnFileName the name for the compdef definition files
 #' @param dataFileName the name for the data files
 #' @param instanceFileName the name for the instance files
+#' @param resultsCacheFileName the name for the instance results cache files for HE computations
 #' @param ssl_verifyhost integer value, usually `1L`, but for
 #'     testing with snake-oil certs, one might set this to `0L`
 #' @param ssl_verifypeer integer value, usually `1L`, but for
@@ -134,6 +135,7 @@ distcompSetup <- function(workspacePath = "",
                           defnFileName = "defn.rds",
                           dataFileName = "data.rds",
                           instanceFileName = "instance.rds",
+                          resultsCacheFileName = "results_cache.rds",
                           ssl_verifyhost = 1L,
                           ssl_verifypeer = 1L) {
   ## TODO: In the next version, this should be stuffed in an R6 class
@@ -175,6 +177,7 @@ distcompSetup <- function(workspacePath = "",
                                   defnFileName = defnFileName,
                                   dataFileName = dataFileName,
                                   instanceFileName = instanceFileName,
+                                  resultsCacheFileName = resultsCacheFileName,
                                   sslConfig = config(ssl_verifyhost = ssl_verifyhost,
                                     ssl_verifypeer = ssl_verifypeer))
   distcompEnv[["computationInfo"]] <- list()
@@ -216,44 +219,56 @@ getConfig <- function(...) {
 
 #' Make a worker object given a definition and data
 #' @description The function `makeWorker` returns an object of the
-#' appropriate type based on a computation definition and sets the data for
-#' the object. The types of objects that can be created depend upon the
-#' available computations
+#'     appropriate type based on a computation definition and sets the
+#'     data for the object. The types of objects that can be created
+#'     depend upon the available computations
 #' @seealso [availableComputations()]
 #' @param defn the computation definition
 #' @param data the data for the computation
-#' @return a worker object of the appropriate class based on the definition
+#' @param pubkey_bits the number of bits for the public key (used only
+#'     if `he` is `TRUE` in computation definition)
+#' @param pubkey_n the `n` for public key (used only if `he` is `TRUE`
+#'     in computation definition)
+#' @param den_bits the number of bits for the denominator (used only
+#'     if `he` is `TRUE` in computation definition)
+#' @return a worker object of the appropriate class based on the
+#'     definition
 #'
 #' @export
-makeWorker <- function (defn, data) {
+makeWorker <- function (defn, data, pubkey_bits = NULL, pubkey_n = NULL, den_bits = NULL) {
   compType <- defn$compType
   available <- availableComputations()
   k <- match(compType, names(available))
   if (is.na(k)) {
     stop(sprintf("No such computation: %s", compType))
   } else {
-    available[[k]]$makeWorker(defn, data)
+      available[[k]]$makeWorker(defn = defn, data = data, pubkey_bits = pubkey_bits,
+                                pubkey_n = pubkey_n, den_bits = den_bits)
   }
 }
 
 #' Make a master object given a definition
 #' @description The function `makeMaster` returns a master object
-#' corresponding to the definition. The types of master objects that can
-#' be created depend upon the available computations
+#'     corresponding to the definition. The types of master objects
+#'     that can be created depend upon the available computations
 #' @seealso [availableComputations()]
 #' @param defn the computation definition
-#' @return a master object of the appropriate class based on the definition
+#' @param partyNumber the number of the noncooperating party, which
+#'     can be optionally set if HE is desired
+#' @param debug a debug flag
+#' @return a master object of the appropriate class based on the
+#'     definition
 #'
 #' @export
-makeMaster <- function(defn) {
-  compType <- defn$compType
-  available <- availableComputations()
-  k <- match(compType, names(available))
-  if (is.na(k)) {
-    stop(sprintf("No such computation: %s", compType))
-  } else {
-    available[[k]]$makeMaster(defn)
-  }
+makeMaster <- function(defn, partyNumber = NULL, debug = FALSE) {
+    compType <- defn$compType
+    available <- availableComputations()
+    k <- match(compType, names(available))
+    if (is.na(k)) {
+        stop(sprintf("No such computation: %s", compType))
+    } else {
+        available[[k]]$makeMaster(defn = defn, partyNumber = partyNumber, debug = debug)
+    }
 }
 
 #' Return the currently available (implemented) computations
@@ -301,11 +316,24 @@ availableComputations <- function() {
                          compType = getComputationInfo("compType"),
                          projectName = getComputationInfo("projectName"),
                          projectDesc = getComputationInfo("projectDesc"),
+                         he = getComputationInfo("he"),
                          filterCondition = getComputationInfo("filterCondition"),
                          stringsAsFactors=FALSE)
           },
-          makeMaster = function(defn, debug = FALSE) QueryCountMaster$new(defn = defn, debug = debug),
-          makeWorker = function(defn, data) QueryCountWorker$new(defn = defn, data = data)
+          makeMaster = function(defn, partyNumber, debug = FALSE) {
+              if(!is.null(defn$he) && defn$he) {
+                  HEQueryCountMaster$new(defn = defn, partyNumber = partyNumber, debug = debug)
+              } else {
+                  QueryCountMaster$new(defn = defn, debug = debug)
+              }
+          },
+          makeWorker = function(defn, data, pubkey_bits = NULL, pubkey_n = NULL, den_bits = NULL) {
+              if (defn$he) {
+                  HEQueryCountWorker$new(defn = defn, data = data, pubkey_bits = pubkey_bits, pubkey_n = pubkey_n, den_bits = den_bits)
+              } else {
+                  QueryCountWorker$new(defn = defn, data = data)
+              }
+          }
       ),
       StratifiedCoxModel = list(
           desc = "Stratified Cox Model",
@@ -317,11 +345,24 @@ availableComputations <- function() {
                          compType = getComputationInfo("compType"),
                          projectName = getComputationInfo("projectName"),
                          projectDesc = getComputationInfo("projectDesc"),
+                         he = getComputationInfo("he"),
                          formula = getComputationInfo("formula"),
                          stringsAsFactors=FALSE)
           },
-          makeMaster = function(defn, debug = FALSE) CoxMaster$new(defn = defn, debug = debug),
-          makeWorker = function(defn, data) CoxWorker$new(defn = defn, data = data)
+          makeMaster = function(defn, partyNumber, debug = FALSE) {
+              if(!is.null(defn$he) && defn$he) {
+                  stop("Not implemented")
+              } else {
+                  CoxMaster$new(defn = defn, debug = debug)
+              }
+          },
+          makeWorker = function(defn, data, pubkey_bits = NULL, pubkey_n = NULL, den_bits = NULL) {
+              if(!is.null(defn$he) && defn$he) {
+                  stop("Not implemented")
+              } else {
+                  CoxWorker$new(defn = defn, data = data)
+              }
+          }
       ),
       RankKSVD = list(
           desc = "Rank K SVD",
@@ -333,15 +374,83 @@ availableComputations <- function() {
                          compType = getComputationInfo("compType"),
                          projectName = getComputationInfo("projectName"),
                          projectDesc = getComputationInfo("projectDesc"),
+                         he = getComputationInfo("he"),
                          rank = getComputationInfo("rank"),
                          ncol = getComputationInfo("ncol"),
                          stringsAsFactors=FALSE)
           },
-          makeMaster = function(defn, debug = FALSE) SVDMaster$new(defn = defn, debug = debug),
-          makeWorker = function(defn, data) SVDWorker$new(defn = defn, data = data)
+          makeMaster = function(defn, partyNumber, debug = FALSE) {
+              if(!is.null(defn$he) && defn$he) {
+                  stop("Not implemented")
+              } else {
+                  SVDMaster$new(defn = defn, debug = debug)
+              }
+          },
+          makeWorker = function(defn, data, pubkey_bits = NULL, pubkey_n = NULL, den_bits = NULL) {
+              if(!is.null(defn$he) && defn$he) {
+                  stop("Not implemented")
+              } else {
+                  SVDWorker$new(defn = defn, data = data)
+              }
+          }
       )
   )
 }
+
+#' Given the definition identifier of an object, instantiate and store
+#' object in workspace
+#' @description This function uses an identifier (`defnId`) to locate
+#'     a stored definition in the workspace to create the appropriate
+#'     object instance. The instantiated object is assigned the
+#'     instanceId and saved under the dataFileName if the latter is
+#'     not `NULL`.  This instantiated object may change state between
+#'     iterations when a computation executes
+#' @param name identifying the NC party
+#' @param ncpId the id indicating the NCP definition
+#' @param instanceId an indentifier to use for the created instance
+#' @param pubkey_bits the public key number of bits
+#' @param pubkey_n the pubkey n
+#' @param den_bits the denominator number of bits for for rational
+#'     approximations
+#' @param dataFileName a file name to use for saving the
+#'     data. Typically `NULL`, this is only needed when one is using a
+#'     single opencpu server to behave like multiple sites in which
+#'     case the data file name serves to distinguish the site-specific
+#'     data files.  When it is `NULL`, the data file name is taken
+#'     from the configuration settings
+#' @import utils
+#' @return TRUE if everything goes well
+#' @export
+createNCPInstance <- function (name, ncpId, instanceId, pubkey_bits, pubkey_n, den_bits, dataFileName = NULL) {
+    config <- getConfig()
+    defnPath <- paste(config$defnPath, ncpId, sep=.Platform$file.sep)
+    ncpDefnFileName <- paste(name, "defn.rds", sep = "-")
+    ncpDefn <- readRDS(paste(defnPath, ncpDefnFileName, sep=.Platform$file.sep))
+    compDef <- readRDS(paste(defnPath, config$defnFileName, sep=.Platform$file.sep))
+    ncp <- makeNCP(ncp_defn = ncpDefn, comp_defn = compDef, pubkey_bits = pubkey_bits,
+                   pubkey_n = gmp::as.bigz(pubkey_n), den_bits = den_bits)
+
+    ncpDataFileName  <- if (is.null(dataFileName)) {
+                            paste(name, config$dataFileName, sep = "-")
+                        } else {
+                            dataFileName
+                        }
+    sites  <- readRDS(paste(defnPath, ncpDataFileName, sep=.Platform$file.sep))
+    site_names  <- sites$name
+    site_urls  <- sites$url
+    ## The JSON serialization makes the columns as lists, so use double bracket indexing!
+    for (i in seq_along(site_names)) {
+        ncp$addSite(name = site_names[[i]], url = site_urls[[i]])
+    }
+
+    ## Check if the instance folder exists
+    thisInstancePath <- paste(config$instancePath, instanceId, sep=.Platform$file.sep)
+    dir.create(thisInstancePath)
+    ## Save it under the instance id to find it.
+    saveRDS(ncp, file=paste(thisInstancePath, config$instanceFileName, sep=.Platform$file.sep))
+    TRUE
+}
+
 
 #' Return currently implemented data sources
 #' @description The function `availableDataSources` returns the
@@ -387,14 +496,15 @@ makeDefinition <- function(compType) {
   }
 }
 
-#' Given the id of a serialized object, invoke a method on the object with arguments
-#' @description The function `executeMethod` is really the heart of distcomp.
-#' It executes an arbitrary method on an object that has been serialized to the
-#' distcomp workspace with any specified arguments. The result, which is dependent
-#' on the computation that is executed, is returned. If the object needs to save
-#' state between iterations on it, it is automatically serialized back for the ensuing
-#' iterations
-#'
+#' Given the id of a serialized object, invoke a method on the object
+#' with arguments
+#' @description The function `executeMethod` is really the heart of
+#'     distcomp.  It executes an arbitrary method on an object that
+#'     has been serialized to the distcomp workspace with any
+#'     specified arguments. The result, which is dependent on the
+#'     computation that is executed, is returned. If the object needs
+#'     to save state between iterations on it, it is automatically
+#'     serialized back for the ensuing iterations
 #' @param objectId the (instance) identifier of the object on which to invoke a method
 #' @param method the name of the method to invoke
 #' @param ... further arguments as appropriate for the method
@@ -410,6 +520,33 @@ executeMethod <- function(objectId, method, ...) {
     saveRDS(object, file=filePath)
   }
   result
+}
+
+#' Given the id of a serialized object, invoke a method on the object
+#' with arguments using homomorphic encryption
+#' @description The function `executeHEMethod` is a homomorphic
+#'     encryption wrapper around `executeMethod`. It ensures any
+#'     returned result is encrypted using the homomorphic encryption
+#'     function.
+#' @param objectId the (instance) identifier of the object on which to
+#'     invoke a method
+#' @param method the name of the method to invoke
+#' @param ... further arguments as appropriate for the method
+#' @return a list containing an integer and a fractional result converted to characters
+#' @export
+executeHEMethod <- function(objectId, method, ...) {
+  config <- getConfig()
+  filePath <- paste(config$instancePath, objectId, config$instanceFileName, sep=.Platform$file.sep)
+  object <- readRDS(file=filePath)
+  call <- substitute(object$METHOD(...), list(METHOD = as.name(method)))
+  result <- eval(call)
+  if (object$getStateful()) {
+    saveRDS(object, file=filePath)
+  }
+  ## For an HE method, return results as strings since we need to serialize things over
+  ## the wire!
+  list(int = as.character(result$int),
+       frac = as.character(result$frac))
 }
 
 #' Deserialize the result of a http response
@@ -430,22 +567,26 @@ executeMethod <- function(objectId, method, ...) {
 #' @keywords internal
 ## Deserialize URL result, only handle JSON at the moment
 .deSerialize <- function(q) {
-  cType <- headers(q)['content-type']
-  if (cType == "application/json")
-    fromJSON(rawToChar(q$content))
-  else
-    q$content
+    httr::stop_for_status(q)
+    cType <- headers(q)['content-type']
+    if (cType == "application/json")
+        jsonlite::fromJSON(rawToChar(q$content), simplifyDataFrame = FALSE)
+    else
+        q$content
 }
 
 
 #' Given the definition identifier of an object, instantiate and store object in workspace
-#' @description The function `createInstanceObject` uses a definition identified by
+#' @description The function `createWorkerInstance` uses a definition identified by
 #' defnId to create the appropriate object instance. The instantiated object is assigned
 #' the instanceId and saved under the dataFileName if the latter is specified.
 #' This instantiated object may change state between iterations when a computation executes
 #' @seealso [availableComputations()]
 #' @param defnId the identifier of an already defined computation
 #' @param instanceId an indentifier to use for the created instance
+#' @param pubkey_bits number of bits for public key
+#' @param pubkey_n the `n` for public key
+#' @param den_bits the number of bits for the denominator
 #' @param dataFileName a file name to use for saving the data. Typically `NULL`, this
 #' is only needed when one is using a single opencpu server to behave like multiple
 #' sites in which case the data file name serves to distinguish the site-specific data files.
@@ -453,19 +594,19 @@ executeMethod <- function(objectId, method, ...) {
 #' @import utils
 #' @return TRUE if everything goes well
 #' @export
-createInstanceObject <- function (defnId, instanceId, dataFileName=NULL) {
+createWorkerInstance <- function (defnId, instanceId, pubkey_bits = NULL, pubkey_n = NULL, den_bits = NULL, dataFileName=NULL) {
   config <- getConfig()
   defn <- readRDS(paste(config$defnPath, defnId, config$defnFileName, sep=.Platform$file.sep))
   compType <- defn$compType
   available <- availableComputations()
   if (!(compType %in% names(available))) {
-    stop(paste("createInstanceObject: No such computation:", compType))
+    stop(paste("createWorkerInstance: No such computation:", compType))
   }
   data <- readRDS(paste(config$defnPath, defnId,
                         if (is.null(dataFileName)) config$dataFileName else dataFileName,
                         sep=.Platform$file.sep))
 
-  object <- makeWorker(defn, data)
+  object <- makeWorker(defn = defn, data = data, pubkey_bits = pubkey_bits, pubkey_n = pubkey_n, den_bits = den_bits)
 
   ## Check if the instance folder exists
   thisInstancePath <- paste(config$instancePath, instanceId, sep=.Platform$file.sep)
@@ -475,12 +616,64 @@ createInstanceObject <- function (defnId, instanceId, dataFileName=NULL) {
   TRUE
 }
 
+#' Given the definition identifier of an object, instantiate and store
+#' object in workspace
+#' @description The function `createHEWorkerInstance` uses a
+#'     definition identified by defnId to create the appropriate
+#'     object instance for HE computations. The instantiated object is
+#'     searched for in the instance path and loaded if already
+#'     present, otherwise it is created and assigned the instanceId
+#'     and saved under the dataFileName if the latter is specified.
+#'     This instantiated object may change state between iterations
+#'     when a computation executes
+#' @seealso [availableComputations()]
+#' @param defnId the identifier of an already defined computation
+#' @param instanceId an indentifier to use for the created instance
+#' @param pubkey_bits number of bits for public key
+#' @param pubkey_n the `n` for public key
+#' @param den_bits the number of bits for the denominator
+#' @param dataFileName a file name to use for saving the
+#'     data. Typically `NULL`, this is only needed when one is using a
+#'     single opencpu server to behave like multiple sites in which
+#'     case the data file name serves to distinguish the site-specific
+#'     data files.  When it is `NULL`, the data file name is taken
+#'     from the configuration settings
+#' @import utils
+#' @return TRUE if everything goes well
+#' @export
+createHEWorkerInstance <- function (defnId, instanceId, pubkey_bits = NULL, pubkey_n = NULL, den_bits = NULL, dataFileName=NULL) {
+  config <- getConfig()
+  thisInstancePath <- paste(config$instancePath, instanceId, sep=.Platform$file.sep)
+  instanceFile  <- file.path(thisInstancePath, config$instanceFileName)
+  if (!file.exists(instanceFile)) {
+      defn <- readRDS(paste(config$defnPath, defnId, config$defnFileName, sep=.Platform$file.sep))
+      compType <- defn$compType
+      available <- availableComputations()
+      if (!(compType %in% names(available))) {
+          stop(paste("createWorkerInstance: No such computation:", compType))
+      }
+      data <- readRDS(paste(config$defnPath, defnId,
+                            if (is.null(dataFileName)) config$dataFileName else dataFileName,
+                            sep=.Platform$file.sep))
+
+      object <- makeWorker(defn = defn, data = data, pubkey_bits = pubkey_bits, pubkey_n = pubkey_n, den_bits = den_bits)
+      ## Check if the instance folder exists
+      thisInstancePath <- paste(config$instancePath, instanceId, sep=.Platform$file.sep)
+      dir.create(thisInstancePath)
+      ## Save it under the instance id to find it.
+      saveRDS(object, file=paste(thisInstancePath, config$instanceFileName, sep=.Platform$file.sep))
+  }
+  TRUE
+}
+
+
+
 #' Destroy an instance object given its identifier
 #' @description The function `destroyInstanceObject` deletes an object associated
 #' with the instanceId. This is typically done after a computation completes and results
 #' have been obtained.
 #' @param instanceId the id of the object to destroy
-#' @seealso [createInstanceObject()]
+#' @seealso [createWorkerInstance()]
 #' @import utils
 #' @return TRUE if everything goes well
 #' @export
@@ -500,7 +693,7 @@ destroyInstanceObject <- function (instanceId) {
 #' This function is invoked (maybe remotely) on the opencpu server by
 #' [uploadNewComputation()] when a worker site is being set up
 #' @seealso [uploadNewComputation()]
-#' @param defn the identifier of an already defined computation
+#' @param defn an already defined computation
 #' @param data the (local) data to use
 #' @param dataFileName a file name to use for saving the data. Typically `NULL`, this
 #' is only needed when one is using a single opencpu server to behave like multiple
@@ -556,7 +749,7 @@ uploadNewComputation <- function(site, defn, data) {
                      list(defn = defn, data = data)
                  }
       q <- POST(.makeOpencpuURL(urlPrefix = site$url, fn = "saveNewComputation"),
-                body = toJSON(payload),
+                body = jsonlite::toJSON(payload),
                 add_headers("Content-Type" = "application/json"),
                 config=getConfig()$sslConfig
                 )
@@ -566,6 +759,90 @@ uploadNewComputation <- function(site, defn, data) {
   }
 }
 
+#' Save an NCP instance, given the sites as associated data and
+#' possibly a data file name to use
+#' @description The function `saveNewNCP` uses the list of sites
+#'     definition to save a new NCP instance. This is
+#'     typically done for every pair of NCPs used in a computation. The function examines the
+#'     computation definition and uses the identifier therein to
+#'     uniquely refer to the computation instance at the site.  This
+#'     function is invoked (maybe remotely) on the opencpu server by
+#'     [uploadNewComputation()] when a worker site is being set up
+#' @seealso [uploadNewNCP()]
+#' @param defn a definition of the ncp
+#' @param data the list of sites with name and url to use
+#' @param dataFileName a file name to use for saving the
+#'     data. Typically `NULL`, this is only needed when one is using a
+#'     single opencpu server to behave like multiple sites in which
+#'     case the data file name serves to distinguish the site-specific
+#'     data files.  When it is `NULL`, the data file name is taken
+#'     from the definition settings
+#' @return TRUE if everything goes well
+#' @export
+saveNewNCP <- function(defn, data, dataFileName = NULL) {
+  config <- getConfig()
+  defnId <- defn$id
+  browser()
+  thisDefnPath <- paste(config$defnPath, defnId, sep = .Platform$file.sep)
+  ncpDefnFileName <- paste(defn$name, "defn.rds", sep = "-")
+  ## Should tryCatch this
+  if (!file.exists(thisDefnPath)) {
+    dir.create(thisDefnPath)
+  }
+  saveRDS(object = defn, file = paste(thisDefnPath, ncpDefnFileName, sep = .Platform$file.sep))
+  ncpDataFileName  <- if (is.null(dataFileName)) {
+                          paste(defn$name, config$dataFileName, sep = "-")
+                      } else {
+                          dataFileName
+                      }
+  saveRDS(object = data, file = paste(thisDefnPath, ncpDataFileName, sep = .Platform$file.sep))
+  TRUE
+}
+
+#' Upload a new Non-Cooperating Party (NCP) information and sites to
+#' an opencpu server
+#' @description The function `uploadNewNCP` is really a remote version
+#'     of [saveNewNCP()], invoking that function on an opencpu server.
+#'     This is typically done for the two NCPs participating in a
+#'     computation with the list of sites. Note that sites are always
+#'     a list of at least a unique name element (distinguishing the
+#'     site from others) and a url element.
+#' @seealso [saveNewNCP()]
+#' @param defn a definition for the NCP
+#' @param url the url for the NCP. Only one of url and worker can be
+#'     non-null
+#' @param worker the worker for the NCP if local. Only one of url and
+#'     worker can be non-null
+#' @param sites a list of lists, each containing two items, a unique
+#'     `name` and a (not necessarily unique) `url`. This is the data
+#'     for the NCP!
+#' @importFrom stringr str_trim
+#' @importFrom httr POST
+#' @importFrom httr headers
+#' @importFrom httr add_headers
+#' @importFrom jsonlite fromJSON
+#' @importFrom jsonlite toJSON
+#'
+#' @return TRUE if everything goes well
+#' @export
+uploadNewNCP <- function(defn, url = NULL, worker = NULL, sites) {
+    if (is.null(worker)) {
+        localhost <- (grepl("^http://localhost", url) || grepl("^http://127.0.0.1", url))
+        payload <- if (localhost) {
+                       list(defn = defn, data = sites, dataFileName = paste0(defn$name, "-data.rds"))
+                   } else {
+                       list(defn = defn, data = sites)
+                   }
+        q <- POST(.makeOpencpuURL(urlPrefix = url, fn = "saveNewNCP"),
+                  body = jsonlite::toJSON(payload),
+                  add_headers("Content-Type" = "application/json"),
+                  config=getConfig()$sslConfig
+                  )
+        .deSerialize(q)
+    } else {
+
+    }
+}
 
 #' Generate an identifier for an object
 #' @description A hash is generated based on the contents of the object
@@ -755,3 +1032,27 @@ writeCode <- function(defn, sites, outputFilenamePrefix) {
   close(f)
   TRUE
 }
+
+#' Instantiate an noncooperating party
+#' @param ncp_defn the NCP definition
+#' @param comp_defn the computation definition
+#' @param sites a list of sites each entry a named list of name, url, worker
+#' @param pubkey_bits number of bits for public key
+#' @param pubkey_n the n for the public key
+#' @param den_bits the log to base 2 of the denominator
+#' @return an NCP object
+#' @export
+makeNCP  <- function(ncp_defn, comp_defn, sites = list(), pubkey_bits = NULL, pubkey_n = NULL, den_bits = NULL) {
+    NCP$new(ncp_defn = ncp_defn, comp_defn = comp_defn, sites = sites, pubkey_bits = pubkey_bits, pubkey_n = pubkey_n, den_bits = den_bits)
+}
+
+
+#' Instantiate a master process for HE operations
+#' @param defn the computation definition
+#' @return an master object for HE operations
+#' @export
+makeHEMaster  <- function(defn) {
+    HEMaster$new(defn)
+}
+
+